Change score from Float to Integer

+++ db/migrate/20121013000510_create_rankings.rb
@@ -2,7 +2,7 @@ class CreateRankings < ActiveRecord::Migration
   def change
     create_table :rankings do |t|
       t.belongs_to :user
-      t.float :score
+      t.integer :score
 
       t.timestamps
     end
